Update Collectibles View  (#3543)
* Add Open xNFT animation

* Nft right click menu
